+++ Manatee.Trello.IntegrationTests/CardTests.cs
@@ -69,9 +69,10 @@ namespace Manatee.Trello.IntegrationTests
 		}
 
 		[Test]
-		[Ignore("Need to determine how to enable custom fields on the board before this will run.")]
 		public async Task CustomFields()
 		{
+			await TestEnvironment.Current.Board.EnsurePowerUp(new CustomFieldsPowerUp());
+
 			var numberField = await TestEnvironment.Current.Board.CustomFields.Add("NumberField", CustomFieldType.Number);
 			var textField = await TestEnvironment.Current.Board.CustomFields.Add("TextField", CustomFieldType.Text);
 			var dateField = await TestEnvironment.Current.Board.CustomFields.Add("DateField", CustomFieldType.DateTime);
@@ -82,10 +83,13 @@ namespace Manatee.Trello.IntegrationTests
 
 			var card = await TestEnvironment.Current.BuildCard();
 
+			await dropDownField.Refresh();
+
 			var today = DateTime.Today;
 			var two = dropDownField.Options.FirstOrDefault(o => o.Text == "two");
+			Assert.NotNull(two);
 
-			await numberField.SetValueForCard(card, 9);
+			await numberField.SetValueForCard(card, 9.6);
 			await textField.SetValueForCard(card, "text");
 			await dateField.SetValueForCard(card, today);
 			await dropDownField.SetValueForCard(card, two);
@@ -93,7 +97,7 @@ namespace Manatee.Trello.IntegrationTests
 
 			await card.Refresh();
 
-			card.CustomFields.OfType<NumberField>().First().Value.Should().Be(9);
+			card.CustomFields.OfType<NumberField>().First().Value.Should().Be(9.6);
 			card.CustomFields.OfType<TextField>().First().Value.Should().Be("text");
 			card.CustomFields.OfType<DateTimeField>().First().Value.Should().Be(today);
 			card.CustomFields.OfType<DropDownField>().First().Value.Text.Should().Be("two");
@@ -202,6 +206,31 @@ namespace Manatee.Trello.IntegrationTests
 				Card.DownloadedFields &= ~Card.Fields.Comments;
 				TrelloConfiguration.RefreshThrottle = TimeSpan.FromSeconds(5);
 			}
+		}
+
+		[Test]
+		public async Task CreatingWithShortId()
+		{
+			var card = await TestEnvironment.Current.BuildCard();
+			var shortId = card.ShortUrl.Split('/').Last();
+
+			TrelloConfiguration.Cache.Remove(card);
+
+			var otherCard = TestEnvironment.Current.Factory.Card(shortId);
+
+			await card.Refresh();
+
+			Assert.AreEqual(card.Id, otherCard.Id);
+		}
+
+		[Test]
+		public async Task CanCopy()
+		{
+			var card = await TestEnvironment.Current.BuildCard();
+			var otherCard = await card.List.Cards.Add(card);
+
+			Assert.AreNotEqual(card.Id, otherCard.Id);
+			Assert.AreEqual(card.Name, otherCard.Name);
 		}
 	}
 }