+++ Manatee.Trello/Internal/Licensing/LicenseHelpers.cs
@@ -4,10 +4,13 @@
 // Modified for use in Manatee.Trello
 #endregion
 
+//#define PERSIST_USAGE
+
 using System;
 using System.Globalization;
 using System.IO;
 using System.Reflection;
+using System.Runtime.InteropServices;
 using System.Threading;
 using Manatee.Json;
 using Manatee.Json.Serialization;
@@ -25,7 +28,33 @@ namespace Manatee.Trello.Internal.Licensing
 
 		public const string BuyMeText = "Please see https://github.com/gregsdennis/Manatee.Trello/wiki/Licensing-3 for information on purchasing a license.";
 
-		private const string DetailsPath = "Manatee.Trello.run";
+		private const string WindowsDetailsPath = @"%LOCALAPPDATA%\Manatee.Trello\Manatee.Trello.run";
+		private const string MacOsDetailsPath = @"~/Library/Manatee.Trello/Manatee.Trello.run";
+		private const string UnixDetailsPath = @"~/.config/Manatee.Trello/Manatee.Trello.run";
+
+		private static string DetailsPath
+		{
+			get
+			{
+#if NET45
+				switch (Environment.OSVersion.Platform)
+					{
+						case PlatformID.MacOSX:
+							return MacOsDetailsPath;
+						case PlatformID.Unix:
+							return UnixDetailsPath;
+						default:
+							return Environment.ExpandEnvironmentVariables(WindowsDetailsPath);
+					}
+#else
+				if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
+					return MacOsDetailsPath;
+				if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+					return UnixDetailsPath;
+				return Environment.ExpandEnvironmentVariables(WindowsDetailsPath);
+#endif
+			}
+		}
 
 		private static readonly object Lock;
 		private static readonly JsonSerializer Serializer;
@@ -60,6 +89,7 @@ namespace Manatee.Trello.Internal.Licensing
 
 		public static void SaveCurrentState()
 		{
+#if PERSIST_USAGE
 			var newDetails = new RunDetails
 				{
 					Retrievals = _retrievalCount,
@@ -69,10 +99,12 @@ namespace Manatee.Trello.Internal.Licensing
 
 			var json = Serializer.Serialize(newDetails);
 			File.WriteAllText(DetailsPath, json.ToString());
+#endif
 		}
 
 		private static void LoadCurrentState()
 		{
+#if PERSIST_USAGE
 			if (!File.Exists(DetailsPath)) return;
 
 			var text = File.ReadAllText(DetailsPath);
@@ -84,6 +116,7 @@ namespace Manatee.Trello.Internal.Licensing
 			_retrievalCount = details.Retrievals;
 			_submissionCount = details.Submissions;
 			_sessionExpiry = Math.Max(0, (long) (details.SessionExpiry - DateTime.Now).TotalMilliseconds);
+#endif
 		}
 
 		public static void IncrementAndCheckRetrieveCount()