+++ Manatee.Trello/ReadOnlyCheckListCollection.cs
@@ -35,7 +35,7 @@ namespace Manatee.Trello
 		public sealed override async Task Refresh(CancellationToken ct = default(CancellationToken))
 		{
 			var endpoint = EndpointFactory.Build(EntityRequestType.Card_Read_CheckLists, new Dictionary<string, object> {{"_id", OwnerId}});
-			var newData = await JsonRepository.Execute<List<IJsonCheckList>>(Auth, endpoint, ct);
+			var newData = await JsonRepository.Execute<List<IJsonCheckList>>(Auth, endpoint, ct, AdditionalParameters);
 
 			Items.Clear();
 			Items.AddRange(newData.Select(jc =>