+++ Manatee.Trello/ReadOnlyOrganizationMembershipCollection.cs
@@ -14,12 +14,10 @@ namespace Manatee.Trello
 	/// </summary>
 	public class ReadOnlyOrganizationMembershipCollection : ReadOnlyCollection<IOrganizationMembership>, IReadOnlyOrganizationMembershipCollection
 	{
-		private Dictionary<string, object> _additionalParameters;
-
 		internal ReadOnlyOrganizationMembershipCollection(Func<string> getOwnerId, TrelloAuthorization auth)
 			: base(getOwnerId, auth)
 		{
-			_additionalParameters = new Dictionary<string, object> {{"fields", "all"}};
+			AdditionalParameters["fields"] = "all";
 		}
 
 		/// <summary>
@@ -48,13 +46,11 @@ namespace Manatee.Trello
 		/// <param name="filters">The filter values.</param>
 		public void Filter(IEnumerable<MembershipFilter> filters)
 		{
-			if (_additionalParameters == null)
-				_additionalParameters = new Dictionary<string, object> { { "filter", string.Empty } };
-			var filter = ((string)_additionalParameters["filter"]);
+			var filter = (string)AdditionalParameters["filter"];
 			if (!filter.IsNullOrWhiteSpace())
 				filter += ",";
 			filter += filters.Select(a => a.GetDescription()).Join(",");
-			_additionalParameters["filter"] = filter;
+			AdditionalParameters["filter"] = filter;
 		}
 
 		/// <summary>
@@ -64,7 +60,7 @@ namespace Manatee.Trello
 		public sealed override async Task Refresh(CancellationToken ct = default(CancellationToken))
 		{
 			var endpoint = EndpointFactory.Build(EntityRequestType.Organization_Read_Memberships, new Dictionary<string, object> {{"_id", OwnerId}});
-			var newData = await JsonRepository.Execute<List<IJsonOrganizationMembership>>(Auth, endpoint, ct, _additionalParameters);
+			var newData = await JsonRepository.Execute<List<IJsonOrganizationMembership>>(Auth, endpoint, ct, AdditionalParameters);
 
 			Items.Clear();
 			Items.AddRange(newData.Select(jom =>